Moved aggregate tvl api related methods to DetailedTvlController.ts and generateDetailedTvlApiResponse.ts
